bridge.sol: uint -> uint256: more explicit, clear.
seen in many other solidity projects with high code quality

+++ contracts/bridge.sol
@@ -6,7 +6,7 @@ pragma solidity ^0.4.17;
 library Helpers {
     /// returns whether `array` contains `value`.
     function addressArrayContains(address[] array, address value) internal pure returns (bool) {
-        for (uint i = 0; i < array.length; i++) {
+        for (uint256 i = 0; i < array.length; i++) {
             if (array[i] == value) {
                 return true;
             }
@@ -16,10 +16,10 @@ library Helpers {
 
     // returns the digits of `inputValue` as a string.
     // example: `uintToString(12345678)` returns `"12345678"`
-    function uintToString(uint inputValue) internal pure returns (string) {
+    function uintToString(uint256 inputValue) internal pure returns (string) {
         // figure out the length of the resulting string
-        uint length = 0;
-        uint currentValue = inputValue;
+        uint256 length = 0;
+        uint256 currentValue = inputValue;
         do {
             length++;
             currentValue /= 10;
@@ -27,7 +27,7 @@ library Helpers {
         // allocate enough memory
         bytes memory result = new bytes(length);
         // construct the string backwards
-        uint i = length - 1;
+        uint256 i = length - 1;
         currentValue = inputValue;
         do {
             result[i--] = byte(48 + currentValue % 10);
@@ -40,7 +40,7 @@ library Helpers {
     /// contain `requiredSignatures` distinct correct signatures
     /// where signer is in `allowed_signers`
     /// that signed `message`
-    function hasEnoughValidSignatures(bytes message, uint8[] vs, bytes32[] rs, bytes32[] ss, address[] allowed_signers, uint requiredSignatures) internal pure returns (bool) {
+    function hasEnoughValidSignatures(bytes message, uint8[] vs, bytes32[] rs, bytes32[] ss, address[] allowed_signers, uint256 requiredSignatures) internal pure returns (bool) {
         // not enough signatures
         if (vs.length < requiredSignatures) {
             return false;
@@ -49,7 +49,7 @@ library Helpers {
         var hash = MessageSigning.hashMessage(message);
         var encountered_addresses = new address[](allowed_signers.length);
 
-        for (uint i = 0; i < requiredSignatures; i++) {
+        for (uint256 i = 0; i < requiredSignatures; i++) {
             var recovered_address = ecrecover(hash, vs[i], rs[i], ss[i]);
             // only signatures by addresses in `addresses` are allowed
             if (!addressArrayContains(allowed_signers, recovered_address)) {
@@ -77,7 +77,7 @@ library HelpersTest {
         return Helpers.uintToString(inputValue);
     }
 
-    function hasEnoughValidSignatures(bytes message, uint8[] vs, bytes32[] rs, bytes32[] ss, address[] addresses, uint requiredSignatures) public pure returns (bool) {
+    function hasEnoughValidSignatures(bytes message, uint8[] vs, bytes32[] rs, bytes32[] ss, address[] addresses, uint256 requiredSignatures) public pure returns (bool) {
         return Helpers.hasEnoughValidSignatures(message, vs, rs, ss, addresses, requiredSignatures);
     }
 }
@@ -117,11 +117,11 @@ library MessageSigningTest {
 
 library Message {
     // layout of message :: bytes:
-    // offset  0: 32 bytes :: uint (little endian) - message length
+    // offset  0: 32 bytes :: uint256 (little endian) - message length
     // offset 32: 20 bytes :: address - recipient address
-    // offset 52: 32 bytes :: uint (little endian) - value
+    // offset 52: 32 bytes :: uint256 (little endian) - value
     // offset 84: 32 bytes :: bytes32 - transaction hash
-    // offset 116: 32 bytes :: uint (little endian) - home gas price
+    // offset 116: 32 bytes :: uint256 (little endian) - home gas price
 
     // bytes 1 to 32 are 0 because message length is stored as little endian.
     // mload always reads 32 bytes.
@@ -143,8 +143,8 @@ library Message {
         return recipient;
     }
 
-    function getValue(bytes message) internal pure returns (uint) {
-        uint value;
+    function getValue(bytes message) internal pure returns (uint256) {
+        uint256 value;
         // solium-disable-next-line security/no-inline-assembly
         assembly {
             value := mload(add(message, 52))
@@ -161,8 +161,8 @@ library Message {
         return hash;
     }
 
-    function getHomeGasPrice(bytes message) internal pure returns (uint) {
-        uint gasPrice;
+    function getHomeGasPrice(bytes message) internal pure returns (uint256) {
+        uint256 gasPrice;
         // solium-disable-next-line security/no-inline-assembly
         assembly {
             gasPrice := mload(add(message, 116))
@@ -178,7 +178,7 @@ library MessageTest {
         return Message.getRecipient(message);
     }
 
-    function getValue(bytes message) public pure returns (uint) {
+    function getValue(bytes message) public pure returns (uint256) {
         return Message.getValue(message);
     }
 
@@ -186,7 +186,7 @@ library MessageTest {
         return Message.getTransactionHash(message);
     }
 
-    function getHomeGasPrice(bytes message) public pure returns (uint) {
+    function getHomeGasPrice(bytes message) public pure returns (uint256) {
         return Message.getHomeGasPrice(message);
     }
 }
@@ -196,14 +196,14 @@ contract HomeBridge {
     /// Number of authorities signatures required to withdraw the money.
     ///
     /// Must be lesser than number of authorities.
-    uint public requiredSignatures;
+    uint256 public requiredSignatures;
 
     /// The gas cost of calling `HomeBridge.withdraw`.
     ///
     /// Is subtracted from `value` on withdraw.
     /// recipient pays the relaying authority for withdraw.
     /// this shuts down attacks that exhaust authorities funds on home chain.
-    uint public estimatedGasCostOfWithdraw;
+    uint256 public estimatedGasCostOfWithdraw;
 
     /// Contract authorities.
     address[] public authorities;
@@ -212,16 +212,16 @@ contract HomeBridge {
     mapping (bytes32 => bool) withdraws;
 
     /// Event created on money deposit.
-    event Deposit (address recipient, uint value);
+    event Deposit (address recipient, uint256 value);
 
     /// Event created on money withdraw.
-    event Withdraw (address recipient, uint value);
+    event Withdraw (address recipient, uint256 value);
 
     /// Constructor.
     function HomeBridge(
-        uint requiredSignaturesParam,
+        uint256 requiredSignaturesParam,
         address[] authoritiesParam,
-        uint estimatedGasCostOfWithdrawParam
+        uint256 estimatedGasCostOfWithdrawParam
     ) public
     {
         require(requiredSignaturesParam != 0);
@@ -254,9 +254,9 @@ contract HomeBridge {
         require(Helpers.hasEnoughValidSignatures(message, vs, rs, ss, authorities, requiredSignatures));
 
         address recipient = Message.getRecipient(message);
-        uint value = Message.getValue(message);
+        uint256 value = Message.getValue(message);
         bytes32 hash = Message.getTransactionHash(message);
-        uint homeGasPrice = Message.getHomeGasPrice(message);
+        uint256 homeGasPrice = Message.getHomeGasPrice(message);
 
         // if the recipient calls `withdraw` they can choose the gas price freely.
         // if anyone else calls `withdraw` they have to use the gas price
@@ -274,10 +274,10 @@ contract HomeBridge {
         // Order of operations below is critical to avoid TheDAO-like re-entry bug
         withdraws[hash] = true;
 
-        uint estimatedWeiCostOfWithdraw = estimatedGasCostOfWithdraw * homeGasPrice;
+        uint256 estimatedWeiCostOfWithdraw = estimatedGasCostOfWithdraw * homeGasPrice;
 
         // charge recipient for relay cost
-        uint valueRemainingAfterSubtractingCost = value - estimatedWeiCostOfWithdraw;
+        uint256 valueRemainingAfterSubtractingCost = value - estimatedWeiCostOfWithdraw;
 
         // pay out recipient
         recipient.transfer(valueRemainingAfterSubtractingCost);
@@ -294,21 +294,21 @@ contract ForeignBridge {
     // following is the part of ForeignBridge that implements an ERC20 token.
     // ERC20 spec: https://github.com/ethereum/EIPs/blob/master/EIPS/eip-20-token-standard.md
 
-    uint public totalSupply;
+    uint256 public totalSupply;
 
     string public name = "ForeignBridge";
 
     /// maps addresses to their token balances
-    mapping (address => uint) public balances;
+    mapping (address => uint256) public balances;
 
     // owner of account approves the transfer of an amount by another account
-    mapping(address => mapping (address => uint)) allowed;
+    mapping(address => mapping (address => uint256)) allowed;
 
     /// Event created on money transfer
-    event Transfer(address indexed from, address indexed to, uint tokens);
+    event Transfer(address indexed from, address indexed to, uint256 tokens);
 
     // returns the ERC20 token balance of the given address
-    function balanceOf(address tokenOwner) public view returns (uint) {
+    function balanceOf(address tokenOwner) public view returns (uint256) {
         return balances[tokenOwner];
     }
 
@@ -316,7 +316,7 @@ contract ForeignBridge {
     ///
     /// does not affect `home` chain. does not do a relay.
     /// as specificed in ERC20 this doesn't fail if tokens == 0.
-    function transfer(address recipient, uint tokens) public returns (bool) {
+    function transfer(address recipient, uint256 tokens) public returns (bool) {
         require(balances[msg.sender] >= tokens);
         // fails if there is an overflow
         require(balances[recipient] + tokens >= balances[recipient]);
@@ -333,11 +333,11 @@ contract ForeignBridge {
 
     // created when `approve` is executed to mark that
     // `tokenOwner` has approved `spender` to spend `tokens` of his tokens
-    event Approval(address indexed tokenOwner, address indexed spender, uint tokens);
+    event Approval(address indexed tokenOwner, address indexed spender, uint256 tokens);
 
     // allow `spender` to withdraw from your account, multiple times, up to the `tokens` amount.
     // calling this function repeatedly overwrites the current allowance.
-    function approve(address spender, uint tokens) public returns (bool) {
+    function approve(address spender, uint256 tokens) public returns (bool) {
         allowed[msg.sender][spender] = tokens;
         Approval(msg.sender, spender, tokens);
         return true;
@@ -380,9 +380,9 @@ contract ForeignBridge {
     /// Number of authorities signatures required to withdraw the money.
     ///
     /// Must be less than number of authorities.
-    uint public requiredSignatures;
+    uint256 public requiredSignatures;
 
-    uint public estimatedGasCostOfWithdraw;
+    uint256 public estimatedGasCostOfWithdraw;
 
     /// Contract authorities.
     address[] public authorities;
@@ -394,18 +394,18 @@ contract ForeignBridge {
     mapping (bytes32 => SignaturesCollection) signatures;
 
     /// triggered when relay of deposit from HomeBridge is complete
-    event Deposit(address recipient, uint value);
+    event Deposit(address recipient, uint256 value);
 
     /// Event created on money withdraw.
-    event Withdraw(address recipient, uint value, uint homeGasPrice);
+    event Withdraw(address recipient, uint256 value, uint256 homeGasPrice);
 
     /// Collected signatures which should be relayed to home chain.
     event CollectedSignatures(address authorityResponsibleForRelay, bytes32 messageHash);
 
     function ForeignBridge(
-        uint _requiredSignatures,
+        uint256 _requiredSignatures,
         address[] _authorities,
-        uint _estimatedGasCostOfWithdraw
+        uint256 _estimatedGasCostOfWithdraw
     ) public
     {
         require(_requiredSignatures != 0);
@@ -424,9 +424,9 @@ contract ForeignBridge {
     /// Used to deposit money to the contract.
     ///
     /// deposit recipient (bytes20)
-    /// deposit value (uint)
+    /// deposit value (uint256)
     /// mainnet transaction hash (bytes32) // to avoid transaction duplication
-    function deposit(address recipient, uint value, bytes32 transactionHash) public onlyAuthority() {
+    function deposit(address recipient, uint256 value, bytes32 transactionHash) public onlyAuthority() {
         // Protection from misbehaving authority
         var hash = keccak256(recipient, value, transactionHash);
 
@@ -456,12 +456,12 @@ contract ForeignBridge {
     /// once `requiredSignatures` are collected a `CollectedSignatures` event will be emitted.
     /// an authority will pick up `CollectedSignatures` an call `HomeBridge.withdraw`
     /// which transfers `value - relayCost` to `recipient` completing the transfer.
-    function transferHomeViaRelay(address recipient, uint value, uint homeGasPrice) public {
+    function transferHomeViaRelay(address recipient, uint256 value, uint256 homeGasPrice) public {
         require(balances[msg.sender] >= value);
         // don't allow 0 value transfers to home
         require(value > 0);
 
-        uint estimatedWeiCostOfWithdraw = estimatedGasCostOfWithdraw * homeGasPrice;
+        uint256 estimatedWeiCostOfWithdraw = estimatedGasCostOfWithdraw * homeGasPrice;
         require(value > estimatedWeiCostOfWithdraw);
 
         balances[msg.sender] -= value;
@@ -480,7 +480,7 @@ contract ForeignBridge {
     ///
     /// for withdraw message contains:
     /// withdrawal recipient (bytes20)
-    /// withdrawal value (uint)
+    /// withdrawal value (uint256)
     /// foreign transaction hash (bytes32) // to avoid transaction duplication
     function submitSignature(bytes signature, bytes message) public onlyAuthority() {
         // ensure that `signature` is really `message` signed by `msg.sender`
@@ -502,7 +502,7 @@ contract ForeignBridge {
     }
 
     /// Get signature
-    function signature(bytes32 hash, uint index) public view returns (bytes) {
+    function signature(bytes32 hash, uint256 index) public view returns (bytes) {
         return signatures[hash].signatures[index];
     }
